Move the bind polyfill to root
The problem with it being in the setup folder was that it didn't run first
before all the other files that required Function.prototype.bind i.e. React.
Even added as the first entry point in the bundle didn't solve the problem.

Include it in the karma config file and add a comment stating that it's only
necessary for PhantomJS.