Fail coverage if files are never required in tests
Previously, if tests were missing for a component, nyc would not generate
coverage for it. This is because the babel plugin was doing the instrumentation
and it obviously never got called on the component that was never required.
Using
https://github.com/istanbuljs/babel-plugin-istanbul/issues/67#issuecomment-257214332
caused nyc to generate coverage for all the files included in nycrc.