Don't load static assets in unit tests
This enables the tests to be run with Mocha in Node without having to bundle
them with Webpack first. This is now consistent with the debug build which uses
Webpack and noop-loader.

The babelrc config for the `tests` env has been split in 2 to not generate
coverage reports when watching the files but still ignore static assets.

The Babel plugin includes an ignore pragma for Istanbul because of a bug causes
it to not be excluded via nycrc.

https://github.com/istanbuljs/nyc/issues/581